#951: Move the filters helper to the division_fixtures directory
