+++ CourageScores/ClientApp/src/components/division_fixtures/filters.test.ts
@@ -10,14 +10,14 @@
     isNextFixtureAfterToday,
     optionallyInvertFilter
 } from "./filters";
-import {IFilter} from "../components/division_fixtures/IFilter";
-import {divisionFixtureBuilder, fixtureDateBuilder, noteBuilder} from "./builders/divisions";
-import {ITournamentSideBuilder, tournamentBuilder} from "./builders/tournaments";
-import {teamBuilder} from "./builders/teams";
-import {DivisionFixtureDateDto} from "../interfaces/models/dtos/Division/DivisionFixtureDateDto";
-import {Filter, NullFilter} from "../components/division_fixtures/Filter";
-import {IEditableDivisionFixtureDateDto} from "../components/division_fixtures/IEditableDivisionFixtureDateDto";
-import {fixtureBuilder} from "./builders/games";
+import {IFilter} from "./IFilter";
+import {divisionFixtureBuilder, fixtureDateBuilder, noteBuilder} from "../../helpers/builders/divisions";
+import {ITournamentSideBuilder, tournamentBuilder} from "../../helpers/builders/tournaments";
+import {teamBuilder} from "../../helpers/builders/teams";
+import {DivisionFixtureDateDto} from "../../interfaces/models/dtos/Division/DivisionFixtureDateDto";
+import {Filter, NullFilter} from "./Filter";
+import {IEditableDivisionFixtureDateDto} from "./IEditableDivisionFixtureDateDto";
+import {fixtureBuilder} from "../../helpers/builders/games";
 
 describe('filters', () => {
     const today = date(0);
