+++ CourageScores/ClientApp/src/components/division_fixtures/filters.ts
@@ -1,15 +1,15 @@
-import {all, any} from "./collections";
-import {AndFilter, Filter, NotFilter, NullFilter, OrFilter} from "../components/division_fixtures/Filter";
-import {isInFuture, isInPast, isDateEqualTo, isToday} from "./dates";
-import {IFilter} from "../components/division_fixtures/IFilter";
+import {all, any} from "../../helpers/collections";
+import {AndFilter, Filter, NotFilter, NullFilter, OrFilter} from "./Filter";
+import {isInFuture, isInPast, isDateEqualTo, isToday} from "../../helpers/dates";
+import {IFilter} from "./IFilter";
 import {
     DivisionTournamentFixtureDetailsDto
-} from "../interfaces/models/dtos/Division/DivisionTournamentFixtureDetailsDto";
-import {DivisionFixtureDto} from "../interfaces/models/dtos/Division/DivisionFixtureDto";
-import {DivisionFixtureDateDto} from "../interfaces/models/dtos/Division/DivisionFixtureDateDto";
-import {FixtureDateNoteDto} from "../interfaces/models/dtos/FixtureDateNoteDto";
-import {IEditableDivisionFixtureDateDto} from "../components/division_fixtures/IEditableDivisionFixtureDateDto";
-import {TournamentSideDto} from "../interfaces/models/dtos/Game/TournamentSideDto";
+} from "../../interfaces/models/dtos/Division/DivisionTournamentFixtureDetailsDto";
+import {DivisionFixtureDto} from "../../interfaces/models/dtos/Division/DivisionFixtureDto";
+import {DivisionFixtureDateDto} from "../../interfaces/models/dtos/Division/DivisionFixtureDateDto";
+import {FixtureDateNoteDto} from "../../interfaces/models/dtos/FixtureDateNoteDto";
+import {IEditableDivisionFixtureDateDto} from "./IEditableDivisionFixtureDateDto";
+import {TournamentSideDto} from "../../interfaces/models/dtos/Game/TournamentSideDto";
 
 export interface IRenderContext {
     lastFixtureDateBeforeToday?: string;
